# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/softwarepkg": {
            "get": {
                "description": "list software packages",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "list software packages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "importer of the softwarePkg",
                        "name": "importer",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "phase of the softwarePkg",
                        "name": "phase",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "count per page",
                        "name": "count_per_page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page num which starts from 1",
                        "name": "page_num",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.SoftwarePkgsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            },
            "post": {
                "description": "apply a new software package",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "apply a new software package",
                "parameters": [
                    {
                        "description": "body of applying a new software package",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.softwarePkgRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        },
        "/v1/softwarepkg/{id}": {
            "get": {
                "description": "get software package",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "get software package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of software package",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.SoftwarePkgReviewDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        },
        "/v1/softwarepkg/{id}/review/abandon": {
            "put": {
                "description": "abandon software package",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "abandon software package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of software package",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        },
        "/v1/softwarepkg/{id}/review/approve": {
            "put": {
                "description": "approve software package",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "approve software package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of software package",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        },
        "/v1/softwarepkg/{id}/review/comment": {
            "post": {
                "description": "create a new software package review comment",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "create a new software package review comment",
                "parameters": [
                    {
                        "description": "body of creating a new software package review comment",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.reviewCommentRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "id of software package",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        },
        "/v1/softwarepkg/{id}/review/comment/{cid}/translate": {
            "post": {
                "description": "translate review comment",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "translate review comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of software package",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cid of review comment",
                        "name": "cid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body of translate review comment",
                        "name": "parm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.translationCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/app.TranslatedReveiwCommentDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        },
        "/v1/softwarepkg/{id}/review/reject": {
            "put": {
                "description": "reject software package",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SoftwarePkg"
                ],
                "summary": "reject software package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of software package",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.SoftwarePkgApplicationDTO": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "license": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "sig": {
                    "type": "string"
                },
                "source_code": {
                    "type": "string"
                }
            }
        },
        "app.SoftwarePkgBasicInfoDTO": {
            "type": "object",
            "properties": {
                "applied_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "importer": {
                    "type": "string"
                },
                "phase": {
                    "type": "string"
                },
                "pkg_name": {
                    "type": "string"
                },
                "repo_link": {
                    "type": "string"
                }
            }
        },
        "app.SoftwarePkgReviewCommentDTO": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "app.SoftwarePkgReviewDTO": {
            "type": "object",
            "properties": {
                "application": {
                    "$ref": "#/definitions/app.SoftwarePkgApplicationDTO"
                },
                "applied_at": {
                    "type": "string"
                },
                "approved_by": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.SoftwarePkgReviewCommentDTO"
                    }
                },
                "id": {
                    "type": "string"
                },
                "importer": {
                    "type": "string"
                },
                "phase": {
                    "type": "string"
                },
                "pkg_name": {
                    "type": "string"
                },
                "rejected_by": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "repo_link": {
                    "type": "string"
                }
            }
        },
        "app.SoftwarePkgsDTO": {
            "type": "object",
            "properties": {
                "pkgs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.SoftwarePkgBasicInfoDTO"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "app.TranslatedReveiwCommentDTO": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "controller.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "controller.reviewCommentRequest": {
            "type": "object",
            "required": [
                "comment"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                }
            }
        },
        "controller.softwarePkgRequest": {
            "type": "object",
            "required": [
                "desc",
                "license",
                "pkg_name",
                "platform",
                "reason",
                "sig",
                "source_code"
            ],
            "properties": {
                "desc": {
                    "type": "string"
                },
                "license": {
                    "type": "string"
                },
                "pkg_name": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "sig": {
                    "type": "string"
                },
                "source_code": {
                    "type": "string"
                }
            }
        },
        "controller.translationCommentRequest": {
            "type": "object",
            "properties": {
                "language": {
                    "type": "string"
                }
            }
        }
    }
}